Move settings help dialogs into settings components folder
