+++ app/(gcforms)/[locale]/(form administration)/form-builder/[id]/settings/components/ResponseDelivery.tsx
@@ -13,9 +13,9 @@ import { Radio } from "@formBuilder/components/shared/MultipleChoice";
 import { Button } from "@clientComponents/globals";
 import { useTemplateStore } from "@lib/store/useTemplateStore";
 import { completeEmailAddressRegex } from "@lib/utils/form-builder";
-import { ResponseDeliveryHelpButton } from "@formBuilder/components/shared/ResponseDeliveryHelpDialog";
-import { FormPurposeHelpButton } from "@formBuilder/components/shared/FormPurposeHelpButton";
-import { ResponseDeliveryHelpButtonWithApi } from "@formBuilder/components/shared/ResponseDeliveryHelpDialogApiWithApi";
+import { ResponseDeliveryHelpButton } from "@formBuilder/[id]/settings/components/ResponseDeliveryHelpDialog";
+import { FormPurposeHelpButton } from "@formBuilder/[id]/settings/components/FormPurposeHelpButton";
+import { ResponseDeliveryHelpButtonWithApi } from "@formBuilder/[id]/settings/components/ResponseDeliveryHelpDialogApiWithApi";
 import {
   ClassificationType,
   ClassificationSelect,
